Fixed deleting new counter behavior.

+++ CountOn/ViewController/DetailViewController.swift
@@ -126,22 +126,30 @@ class DetailViewController: ASViewController<ASDisplayNode> {
         doneCancelBar.deleteButton.rx
             .tap
             .subscribe(onNext: { [weak self] _ in
-                let alert = UIAlertController(title: R.string.localizable.thisCounterIsAboutToBeDeleted(), message: R.string.localizable.thisCannotBeUndoneAreYouSure(), preferredStyle: UIAlertControllerStyle.alert)
-                let cancelAction = UIAlertAction(title: R.string.localizable.cancel(), style: UIAlertActionStyle.cancel, handler: nil)
-                let yesAction = UIAlertAction(title: R.string.localizable.yes(), style: UIAlertActionStyle.destructive) { _ in
-                    self?.detailNode.delete()
-                    self?.dismiss(animated: true, completion: nil)
+                if let isInEditMode = self?.detailNode.isInEditMode, isInEditMode {
+                    let alert = UIAlertController(title: R.string.localizable.thisCounterIsAboutToBeDeleted(), message: R.string.localizable.thisCannotBeUndoneAreYouSure(), preferredStyle: UIAlertControllerStyle.alert)
+                    let cancelAction = UIAlertAction(title: R.string.localizable.cancel(), style: UIAlertActionStyle.cancel, handler: nil)
+                    let yesAction = UIAlertAction(title: R.string.localizable.yes(), style: UIAlertActionStyle.destructive) { _ in
+                        
+                        self?.detailNode.delete()
+                        
+                        self?.dismiss(animated: true, completion: nil)
 //                TODO: Now I clear search field
 //                      to prevent crash when deleteing from the search result.
 //                      But hopefully I do not have to do that
-                    SearchAddBarView.shared.searchField.clear()
+                        SearchAddBarView.shared.searchField.clear()
 //                CounterStore.shared.reset()
-                    
+                        
+                        SearchAddBarView.shared.show()
+                    }
+                    alert.addAction(cancelAction)
+                    alert.addAction(yesAction)
+                    self?.present(alert, animated: true, completion: nil)
+                } else {
+                    self?.dismiss(animated: true, completion: nil)
+                    SearchAddBarView.shared.searchField.clear()
                     SearchAddBarView.shared.show()
                 }
-                alert.addAction(cancelAction)
-                alert.addAction(yesAction)
-                self?.present(alert, animated: true, completion: nil)
             })
             .disposed(by: disposeBag)
